edited the nav bar
changed some names and icons
