+++ app/src/main/java/com/example/pumppatrol/MainActivity.kt
@@ -26,7 +26,7 @@ class MainActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications
+                R.id.navigation_home, R.id.navigation_workout, R.id.navigation_history
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
