+++ app/src/main/java/com/example/pumppatrol/ui/history/HistoryViewModel.kt
@@ -1,0 +1,13 @@
+package com.example.pumppatrol.ui.history
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class HistoryViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is where the user can view their workout history"
+    }
+    val text: LiveData<String> = _text
+}
