+++ app/src/main/java/com/example/pumppatrol/ui/home/HomeViewModel.kt
@@ -7,7 +7,8 @@ import androidx.lifecycle.ViewModel
 class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is the home Page"
     }
+
     val text: LiveData<String> = _text
 }
