+++ app/src/main/res/drawable/outline_book_4_spark_24.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#000000" android:viewportHeight="960" android:viewportWidth="960" android:width="24dp">
+      
+    <path android:fillColor="@android:color/white" android:pathData="M200,647Q210,644 219.5,642Q229,640 240,640L280,640L280,160L240,160Q223,160 211.5,171.5Q200,183 200,200L200,647ZM240,880Q190,880 155,845Q120,810 120,760L120,200Q120,150 155,115Q190,80 240,80L520,80Q520,100 520,120Q520,140 520,160L360,160L360,640L600,640L600,520Q620,520 640,520Q660,520 680,520L680,720L240,720Q223,720 211.5,731.5Q200,743 200,760Q200,777 211.5,788.5Q223,800 240,800L760,800L760,480Q780,480 800,480Q820,480 840,480L840,880L240,880ZM200,647Q200,644 200,642Q200,640 200,640L200,640L200,500Q200,471 200,406Q200,341 200,260Q200,234 200,208.5Q200,183 200,160L200,160Q200,160 200,171.5Q200,183 200,200L200,647ZM700,480Q700,388 764,324Q828,260 920,260Q828,260 764,196Q700,132 700,40Q700,132 636,196Q572,260 480,260Q572,260 636,324Q700,388 700,480Z"/>
+    
+</vector>
